+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.1.0
 * UPD: transfered to API 1.10
-* UPD: added localization
+* UPD: added new localization string
 
 ## 1.0.1
 * Add functionality When field "Values" is absent, we can use weight === 1 for each link